Continuous integration setup (#1)
The what
=======
This commit adds the continuous-integration workflow, which can be found in
the .github\workflows\continuous_integration.yaml file.

This workflow runs a variety of tasks to build and test any changes.

The why
=======
Continuous integration has been set up to ensure that the software
will work on a variety of systems. 

It has been built this way to enable testing of any changes can be
carried out automatically, preventing breaking changes to the main
branch.

This can be added to in the future to allow more complex continuous
integration steps.