Add continuous integration workflow
Changes
---------

This commit adds the continuous-integration workflow, which can be found
in the .github\workflows\continuous_integration.yaml file.

This workflow runs a variety of tasks to build and test any changes.

Reasons
---------

Continuous integration has been set up to ensure that the software
will work on a variety of systems.

It has been built this way to enable the testing of any changes to be
carried out automatically, preventing breaking changes to the main
branch.

This can be added to in the future to allow more complex continuous
integration steps.