Add user input class
Changes
---------

This commit adds the Input base class and CommandLineInput class.

Reasons
---------

The Input base class provides a common interface for interacting with the
user via any source of input.

The first concrete class added based on Input is the the CommandLineInput
class, as the command line is a quick and easy way to get user input.