Add text output class
Changes
---------

This commit adds the Output base class and CommandLineOutput class.

Reasons
---------

The Output base class provides a common interface for providing output
to the user via any source of output.

The first concrete class added based on Output is the the CommandLineOutput
class, as the command line is a quick and easy way to display output.