Align CommandLineInput and CommandLineOutput classes
Changes
---------

This commit aligns the CommandLineInput and CommandLineOutput classes,
and adds parametrised unit tests for them.

Reasons
---------

Aligning the way CommandLineInput and CommandLineOutput displays the
prompt/prefix allows for easier and simpler usage of the classes.

The unit tests for these classes have been converted to make them
easier to maintain.