Refactor unit tests
Changes
---------
This commit moves unit tests for input functionality to test_inputs.py and
splits the parametrised tests for inputs in test_inputs.py.

This commit also moves unit tests for output functionality to test_outputs.py
and splits the parametrised tests for outputs in test_outputs.py

Reasons
---------
These changes improve the code quality throughout.

It additionally makes it easier to maintain and add to the unit tests.