Implement logging
Changes
---------
This commit adds the ability to add logging by creating a custom Logger
class, which wraps the inbuilt Logging module. It also implements this
logger to all the custom modules.

Reasons
---------
These changes improve the quality of code and adds the ability to identify
and diagnose issues quickly, track the flow of execution, and monitor the
performance.