Augment dataset
Changes
---------
This commit adds a new method for creating and generating datasets, and
uses this to create a new intent dataset. It also adds functions for the
augmentation pipeline in `main.py` and adds other functions for creating
and interacting with datasets.

Reasons
---------
This allows faster creation of more robust datasets to create better
artificial intelligence models.